Move fetch-hypothesis-annotation into third-parties #1218
