+++ src/third-parties/hypothesis/index.ts
@@ -1,3 +1,4 @@
+export { fetchHypothesisAnnotation } from './fetch-hypothesis-annotation';
 export { fetchEvaluationsByGroupSince } from './fetch-evaluations-by-group';
 export { fetchEvaluationsByUserSince } from './fetch-evaluations-by-user';
 export { Annotation } from './annotation';
