+++ test/http/server.ts
@@ -13,11 +13,11 @@ import { createRouter } from '../../src/http/router';
 import { createApplicationServer } from '../../src/http/server';
 import { Adapters } from '../../src/infrastructure';
 import { FetchDataset } from '../../src/infrastructure/fetch-dataset';
-import { fetchHypothesisAnnotation } from '../../src/infrastructure/fetch-hypothesis-annotation';
 import { fetchReview } from '../../src/infrastructure/fetch-review';
 import { inMemoryGroupRepository } from '../../src/infrastructure/in-memory-groups';
 import { FetchCrossrefArticle } from '../../src/third-parties/crossref';
 import { fetchDataciteReview } from '../../src/third-parties/datacite';
+import { fetchHypothesisAnnotation } from '../../src/third-parties/hypothesis';
 import * as DE from '../../src/types/data-error';
 import { FollowList } from '../../src/types/follow-list';
 import { SanitisedHtmlFragment } from '../../src/types/sanitised-html-fragment';
